requirements.txt
langchain==0.2.16
langchain-community==0.2.16
langchain-core==0.2.38
langchain-text-splitters==0.2.4
langchain-huggingface==0.0.3
langchain-ollama==0.1.3
qdrant-client==1.10.1
sentence-transformers
pymupdf 
pillow
dotenv